+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## HEAD
+
+* `derivatives` – Add `Attacher.derivatives` alias for `Attacher.derivatives_processor` (@janko)
+
 ## 3.0.1 (2019-10-17) [[release notes]](https://shrinerb.com/docs/release_notes/3.0.1)
 
 * `metadata_attributes` – Fix exception being raised when there is no attached file (@janko)